refactor(ui): now_rfc3339 を ui::timestamp に共通化
QuizUI 内の private 関数だった now_rfc3339 / days_to_ymd を
src/ui/timestamp.rs に切り出して TA25 でも使えるようにする。挙動・出力フォーマット・テストとも不変。
